Attempt to provide metric between unified characters.
This tries to add a metric between unified characters to be used
in a clustering pass. Currently this is blocked because of the way
instances are set up for selecting the metric from the metadata
(there are also unique challenges for the dynamic characters) as
the metric needs to perform directOptimization or similar.