Attempt to fix spaceleaks arising from fromList
This commit adds a function trivialRefDAG for use in the unification stage
when we construct a trivial solution. The current version of fromList has
a spaceleak which this commit does not fix. We do use difference lists
instead of lists when finding the roots as this should perform better and
leak less space than the version using lists.