Attempt to fix space leak in unification
This commit uses the new trivialRefDAG function to try to fix
the spaceleak from trying to construct a trivial solution. We
also make a few other changes (like let-floating the trivial root
name so this result is not re-computed). When constructing the
trivial solution we use difference lists for non-empty lists as this
should leak less space and have better performance.