Move towards final functionality
This commit moves towards the finalizing of the new dynamic character
re-rooting code. It also formats the code into various sections to
add to the readability and provides documentation for functions
and the majority of the major ideas. The code is significantly more
strongly typed than the previous version and so makes use of various
custom types for the different aspects.